Install pyhmy as a Python package instead of a git submodule
Some drawbacks of using Git submodules are:
- Not being able to pin the package version
- We would need to install their dev dependencies to run their code

The pyhmy package being used is a fork of the original project.
This is because the original project pins the version of
eth-account to 0.5.4, which is incompatible with the current
version of web3.py

Using a forked version is not ideal either. But we can limit
the dependencies installed with it.